+++ packages/atomic-angular/projects/atomic-angular/src/lib/stencil-generated/components.ts
@@ -3004,6 +3004,7 @@ export class AtomicCommerceSortDropdown {
   constructor(c: ChangeDetectorRef, el: ElementRef, protected z: NgZone) {
     c.detach();
     this.el = el.nativeElement;
+    
   }
 }
 
@@ -3029,6 +3030,7 @@ export class AtomicIcon {
   constructor(c: ChangeDetectorRef, el: ElementRef, protected z: NgZone) {
     c.detach();
     this.el = el.nativeElement;
+    
   }
 }