+++ packages/atomic/src/components/commerce/index.ts
@@ -1,5 +1,2 @@
-// Export the class of the components migrated to Lit here, like this:
-export {AtomicComponentError} from '../common/atomic-component-error/atomic-component-error';
-export {AtomicIcon} from '../common/atomic-icon/atomic-icon';
-export {AtomicCommerceSortDropdown} from './atomic-commerce-sort-dropdown/atomic-commerce-sort-dropdown';
-export {AtomicCommerceFacet} from '../commerce/facets/atomic-commerce-facet/atomic-commerce-facet';
+// Auto-generated file
+export {AtomicCommerceSortDropdown} from './atomic-commerce-sort-dropdown/atomic-commerce-sort-dropdown.js';