+++ packages/atomic/src/components/common/facets/facet-value-link/stencil-facet-value-link.tsx
@@ -14,6 +14,7 @@ export const FacetValueLink: FunctionalComponent<FacetValueLinkProps> = (
   const ariaLabel = props.i18n.t('facet-value', {
     value: props.displayValue,
     count: props.numberOfResults,
+    formattedCount: count,
     interpolation: {escapeValue: false},
   });