+++ packages/atomic/src/index.ts
@@ -14,14 +14,14 @@ export {PopoverChildFacet} from './components/common/facets/popover/popover-type
 
 export {resultContext} from './components/search/result-template-components/result-template-decorators';
 export {productContext} from './components/commerce/product-template-components/product-template-decorators';
-export {SelectChildProductEventArgs} from './components/commerce/product-template-components/atomic-product-children/atomic-product-children';
+export {SelectChildProductEventArgs} from './components/commerce/product-template-components/atomic-product-children/select-child-product-event';
 export {
   dispatchSearchBoxSuggestionsEvent,
   SearchBoxSuggestionElement,
   SearchBoxSuggestions,
   SearchBoxSuggestionsBindings,
   SearchBoxSuggestionsEvent,
-} from './components/common/suggestions/suggestions-common';
+} from './components/common/suggestions/stencil-suggestions-common';
 export {RedirectionPayload} from './components/search/atomic-search-box/redirection-payload';
 
 export {RecsBindings} from './components/recommendations/atomic-recs-interface/atomic-recs-interface';