+++ packages/atomic/storybook-utils/common/common-meta-parameters.ts
@@ -2,4 +2,7 @@ import {Parameters} from '@storybook/web-components';
 
 export const parameters: Parameters = {
   controls: {expanded: true, hideNoControlsWarning: true},
+  docs: {
+    story: {autoplay: true},
+  },
 };