+++ packages/atomic/vitest.config.ts
@@ -76,14 +76,20 @@ export default defineConfig({
       'src/**/initialization-utils.spec.ts',
       'src/**/search-layout.spec.ts',
     ],
+    restoreMocks: true,
     setupFiles: ['./vitest-utils/setup.ts'],
     globals: true,
+    deps: {
+      moduleDirectories: ['node_modules', path.resolve('../../packages')],
+    },
     browser: {
-      enabled: true,
-      name: 'chromium',
       provider: 'playwright',
-      // https://playwright.dev
-      providerOptions: {},
+      enabled: true,
+      instances: [
+        {
+          browser: 'chromium',
+        },
+      ],
     },
   },
 });