+++ packages/samples/headless-ssr/app-router/src/app/react/page.tsx
@@ -24,12 +24,14 @@ import {NextJsAppRouterNavigatorContext} from '../../navigatorContextProvider';
  * These values will be added to the payload of the search request when the search page is rendered.
  */
 
-export default async function Search(url: {
-  searchParams: {[key: string]: string | string[] | undefined};
+export default async function Search({
+  searchParams,
+}: {
+  searchParams: Promise<{[key: string]: string | string[] | undefined}>;
 }) {
   // Convert URL search parameters into a format that Coveo's search engine can understand.
   const {toSearchParameters} = buildSSRSearchParameterSerializer();
-  const searchParameters = toSearchParameters(url.searchParams);
+  const searchParameters = toSearchParameters(await searchParams);
 
   // Defines hard-coded context values to simulate user-specific information.
   const contextValues = {
@@ -38,7 +40,7 @@ export default async function Search(url: {
   };
 
   // Sets the navigator context provider to use the newly created `navigatorContext` before fetching the app static state
-  const navigatorContext = new NextJsAppRouterNavigatorContext(headers());
+  const navigatorContext = new NextJsAppRouterNavigatorContext(await headers());
 
   setNavigatorContextProvider(() => navigatorContext);